requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14,<4.0
django-filter>=23.0
psycopg2-binary>=2.9
python-dotenv>=1.0
